Example from p 127 coded
Can generate an alternative representation for values now.

If we know 254 is the max number, so we have [0-254], then we can
represent each of those values with the result of:

(n % 3, n % 5, n % 17) where n is [0-254].

The claim from Graham et. al. is that we can perform addition,
subtraction and multiplication on the components of the tuple
independently.

Their example is:

7 * 13 modulu 15.
The format is: (n % 3, n % 5)
7   = (1, 2)
13  = (1, 3)
(1, 2) * (1, 3) = (1*1 % 3, 2*3 % 5) = (1, 1).

(1, 1) is a unique representation for the number 1 here.

This is intersting in of itself. But it is also interesting if it is
sometimes possible to go the other way around.

Imagine having a 2D graph, and that given that you know what ranges
x and y values can be in we could represent any position (x,y) with a
single scalar value. This is ofc trivial if we just assign a strategy in
which order to visit all possible points in the graph. But that might be
too slow and cumbersome if the number of potential solutions is large.

This might give us another way to go about it.

Why is a single number better than a tuple? Well if we have 64bit
numbers we might have an efficient way of encoding all positions in a
single scalar without creating cumbersome bitpatterns yourself.

Or maybe this is just for fun, duno. We will see!