Generalized the snake program a bit
Creating this commit after-the-fact so Im quickly trying to understand
what I was working on.

Snake is no longer hardcoded to the given size, other sizes can be used,
however they should be choosen with care. But once that is done it will
find the "unit scalars" for you.

This commit also contains some new realizations that I apparently had
last time I was tinkering with the code that I did not commit.